Reworked Vuex
