+++ src/components/NavBar.vue
@@ -2,28 +2,28 @@
   <nav class="bg-white px-8 pt-2 shadow-md">
     <div class="flex justify-center">
       <router-link
-        v-if="!currentUser"
+        v-if="!isLoggedIn"
         to="/login"
         class="no-underline text-grey-600 border-b-2 border-transparent uppercase tracking-wide font-bold text-xs py-3 mr-8"
       >
         Login
       </router-link>
       <router-link
-        v-if="!currentUser"
+        v-if="!isLoggedIn"
         to="/register"
         class="no-underline text-grey-600 border-b-2 border-transparent uppercase tracking-wide font-bold text-xs py-3 mr-8"
       >
         Register
       </router-link>
       <router-link
-        v-if="!!currentUser"
+        v-if="isLoggedIn"
         to="/"
         class="no-underline text-grey-600 border-b-2 border-transparent uppercase tracking-wide font-bold text-xs py-3 mr-8"
       >
         Todo List
       </router-link>
       <button
-        v-if="!!currentUser"
+        v-if="isLoggedIn"
         class="bg-transparent uppercase text-xs hover:bg-teal-600 text-teal-700 font-semibold hover:text-white py-1 px-3 mr-8 border border-teal-600 hover:border-transparent rounded"
         @click.prevent="logout"
       >
@@ -34,14 +34,14 @@
 </template>
 
 <script>
-import { mapState } from 'vuex'
+import { mapGetters } from 'vuex'
 export default {
   computed: {
-    ...mapState(['currentUser'])
+    ...mapGetters('users', ['isLoggedIn'])
   },
   methods: {
     logout() {
-      this.$store.dispatch('LOGOUT')
+      this.$store.dispatch('users/logoutUser')
     }
   }
 }
